removed an monitored logs and improve some ui and the logics

+++ src/controllers/data.ts
@@ -6,8 +6,17 @@ import {
   getDoc,
   deleteDoc,
 } from "firebase/firestore";
-import { db } from "../auth/firebaseConfig";
+import { auth, db } from "../auth/firebaseConfig";
 import { JobPost } from "../types/types";
+import { onAuthStateChanged } from "firebase/auth";
+import { authState } from "../auth/auth";
+
+// Capture the authenticated user's UID
+let currentUserUID: string | null = null;
+onAuthStateChanged(auth, (user) => {
+  authState.isAuthenticated = !!user;
+  currentUserUID = user?.uid || null;
+});
 
 // Define the collection reference
 const dbCollection = collection(db, "jobs");
@@ -47,7 +56,16 @@ export async function readMoreDetailsAboutAJobPost(id: string) {
 // Function to add a new job post
 export async function addNewJobPost(inputData: JobPost) {
   try {
-    const docRef = await addDoc(dbCollection, inputData);
+    if (!currentUserUID) {
+      throw new Error("User is not authenticated");
+    }
+
+    const jobPostData = {
+      ...inputData,
+      uid: currentUserUID, // Add UID to the job post data
+    };
+
+    const docRef = await addDoc(dbCollection, jobPostData);
     console.log("Job post created with ID: ", docRef.id);
   } catch (error) {
     console.log("An error occurred when creating a new job post: ", error);
@@ -58,7 +76,22 @@ export async function addNewJobPost(inputData: JobPost) {
 // Function to remove a closed job post
 export async function removeClosedJobPost(id: string) {
   try {
+    if (!currentUserUID) {
+      throw new Error("User is not authenticated");
+    }
+
     const jobPostDoc = doc(dbCollection, id);
+    const jobPostSnapshot = await getDoc(jobPostDoc);
+
+    if (!jobPostSnapshot.exists()) {
+      throw new Error("Job post does not exist");
+    }
+
+    const jobPostData = jobPostSnapshot.data();
+    if (jobPostData.uid !== currentUserUID) {
+      throw new Error("User is not authorized to delete this job post");
+    }
+
     await deleteDoc(jobPostDoc);
     console.log("Job post deleted with ID: ", id);
   } catch (error) {
